refactor test files into ___ folder
